requirements.txt
pillow
requests
pyautogui
jinja2